Use Thread.sleep() instead of toast
The previous implementation caused some issues where when the theme
install was intentionally cancelled, the while loop will run infinitely
until the app crashes.

Another way of doing this is to show a dialog for a set period of time
without using a while loop.

Thank you @DeltaROM (Jacob McSwain)